+++ examples/with-create-react-app/src/App.tsx
@@ -1,11 +1,11 @@
-import { ConnectButton } from "@rainbow-me/rainbowkit";
+import { ConnectButton } from '@rainbow-me/rainbowkit';
 
 const App = () => {
   return (
     <div
       style={{
-        display: "flex",
-        justifyContent: "flex-end",
+        display: 'flex',
+        justifyContent: 'flex-end',
         padding: 12,
       }}
     >