+++ examples/with-next-siwe-next-auth/pages/_app.tsx
@@ -1,18 +1,18 @@
-import "../styles/global.css";
-import "@rainbow-me/rainbowkit/styles.css";
-import type { AppProps } from "next/app";
+import '../styles/global.css';
+import '@rainbow-me/rainbowkit/styles.css';
+import type { AppProps } from 'next/app';
 
 import {
   RainbowKitProvider,
   getDefaultWallets,
   getDefaultConfig,
-} from "@rainbow-me/rainbowkit";
+} from '@rainbow-me/rainbowkit';
 import {
   argentWallet,
   trustWallet,
   ledgerWallet,
-} from "@rainbow-me/rainbowkit/wallets";
-import { WagmiProvider } from "wagmi";
+} from '@rainbow-me/rainbowkit/wallets';
+import { WagmiProvider } from 'wagmi';
 import {
   arbitrum,
   base,
@@ -21,25 +21,25 @@ import {
   polygon,
   sepolia,
   zora,
-} from "wagmi/chains";
+} from 'wagmi/chains';
 
-import { SessionProvider } from "next-auth/react";
-import type { Session } from "next-auth";
+import { SessionProvider } from 'next-auth/react';
+import type { Session } from 'next-auth';
 import {
   RainbowKitSiweNextAuthProvider,
   GetSiweMessageOptions,
-} from "@rainbow-me/rainbowkit-siwe-next-auth";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+} from '@rainbow-me/rainbowkit-siwe-next-auth';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 const { wallets } = getDefaultWallets();
 
 const config = getDefaultConfig({
-  appName: "RainbowKit demo",
-  projectId: "YOUR_PROJECT_ID",
+  appName: 'RainbowKit demo',
+  projectId: 'YOUR_PROJECT_ID',
   wallets: [
     ...wallets,
     {
-      groupName: "Other",
+      groupName: 'Other',
       wallets: [argentWallet, trustWallet, ledgerWallet],
     },
   ],
@@ -50,13 +50,13 @@ const config = getDefaultConfig({
     arbitrum,
     base,
     zora,
-    ...(process.env.NEXT_PUBLIC_ENABLE_TESTNETS === "true" ? [sepolia] : []),
+    ...(process.env.NEXT_PUBLIC_ENABLE_TESTNETS === 'true' ? [sepolia] : []),
   ],
   ssr: true,
 });
 
 const getSiweMessageOptions: GetSiweMessageOptions = () => ({
-  statement: "Sign in to the RainbowKit + SIWE example app",
+  statement: 'Sign in to the RainbowKit + SIWE example app',
 });
 
 const queryClient = new QueryClient();