+++ examples/with-remix/remix.config.js
@@ -2,14 +2,14 @@
  * @type {import('@remix-run/dev').AppConfig}
  */
 module.exports = {
-  ignoredRouteFiles: ["**/.*"],
+  ignoredRouteFiles: ['**/.*'],
   // appDirectory: "app",
   // assetsBuildDirectory: "public/build",
   // serverBuildPath: "build/index.js",
   // publicPath: "/build/",
   serverDependenciesToBundle: [
-    "@rainbow-me/rainbowkit",
-    "@rainbow-me/rainbowkit/wallets",
+    '@rainbow-me/rainbowkit',
+    '@rainbow-me/rainbowkit/wallets',
     /^@?wagmi.*/,
     /.*/,
   ],