+++ examples/with-next-app-i18n/CHANGELOG.md
@@ -1,5 +1,19 @@
 # with-next-app-i18n
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [33a8266]
+  - @rainbow-me/rainbowkit@1.3.6
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [2b0c7b3]
+  - @rainbow-me/rainbowkit@1.3.5
+
 ## 0.0.25
 
 ### Patch Changes