+++ examples/with-next-siwe-next-auth/CHANGELOG.md
@@ -1,5 +1,21 @@
 # with-next-siwe-next-auth
 
+## 0.0.62
+
+### Patch Changes
+
+- Updated dependencies [33a8266]
+  - @rainbow-me/rainbowkit@1.3.6
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.3.4
+
+## 0.0.61
+
+### Patch Changes
+
+- Updated dependencies [2b0c7b3]
+  - @rainbow-me/rainbowkit@1.3.5
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.3.4
+
 ## 0.0.60
 
 ### Patch Changes