+++ examples/with-next-wallet-button/CHANGELOG.md
@@ -1,5 +1,19 @@
 # with-next-wallet-button
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [33a8266]
+  - @rainbow-me/rainbowkit@1.3.6
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [2b0c7b3]
+  - @rainbow-me/rainbowkit@1.3.5
+
 ## 0.0.6
 
 ### Patch Changes