+++ examples/with-next/CHANGELOG.md
@@ -1,5 +1,19 @@
 # with-next
 
+## 0.0.79
+
+### Patch Changes
+
+- Updated dependencies [33a8266]
+  - @rainbow-me/rainbowkit@1.3.6
+
+## 0.0.78
+
+### Patch Changes
+
+- Updated dependencies [2b0c7b3]
+  - @rainbow-me/rainbowkit@1.3.5
+
 ## 0.0.77
 
 ### Patch Changes