+++ examples/with-vite/CHANGELOG.md
@@ -1,5 +1,19 @@
 # with-vite
 
+## 0.0.67
+
+### Patch Changes
+
+- Updated dependencies [33a8266]
+  - @rainbow-me/rainbowkit@1.3.6
+
+## 0.0.66
+
+### Patch Changes
+
+- Updated dependencies [2b0c7b3]
+  - @rainbow-me/rainbowkit@1.3.5
+
 ## 0.0.65
 
 ### Patch Changes