+++ packages/create-rainbowkit/generated-test-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # generated-test-app
 
+## 0.1.81
+
+### Patch Changes
+
+- Updated dependencies [33a8266]
+  - @rainbow-me/rainbowkit@1.3.6
+
+## 0.1.80
+
+### Patch Changes
+
+- Updated dependencies [2b0c7b3]
+  - @rainbow-me/rainbowkit@1.3.5
+
 ## 0.1.79
 
 ### Patch Changes