+++ packages/create-rainbowkit/templates/next-app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # rainbowkit-next-app
 
+## 0.0.81
+
+### Patch Changes
+
+- Updated dependencies [33a8266]
+  - @rainbow-me/rainbowkit@1.3.6
+
+## 0.0.80
+
+### Patch Changes
+
+- Updated dependencies [2b0c7b3]
+  - @rainbow-me/rainbowkit@1.3.5
+
 ## 0.0.79
 
 ### Patch Changes