+++ packages/rainbowkit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rainbow-me/rainbowkit
 
+## 1.3.6
+
+### Patch Changes
+
+- 33a8266: Fixed a bug where `account.displayBalance` for the `<ConnectButton.Custom>` component would sometimes be undefined.
+
+## 1.3.5
+
+### Patch Changes
+
+- 2b0c7b3: Added missing i18n translations for `Wrong network` errors
+
 ## 1.3.4
 
 ### Patch Changes